third_party/WebKit/LayoutTests/fast/viewport/viewport-134-expected.txt

CONSOLE MESSAGE: line 3: The value "device-width;" for key "width" is invalid, and has been ignored.
CONSOLE MESSAGE: line 3: Error parsing a meta element's content: ';' is not a valid key-value pair separator. Please use ',' instead.
ALERT: viewport size 160x176 scale 2 with limits [2, 5] and userScalable true
